Create MultiplyTuples
did something similar for a project I'm working on. Using lambda to call a function with a function as a parameter without calling either function until the code that calls it.

This does something similar but in a simpler manner -- a map function and a simple multiplication function, neither of which the computer unpacks until execution of the main file code.